Consolidate workflow-duration formatter into RunBotCore (#2609)
